Editor: Initial changes following editor.py division in editorstack.py, splitter.py and window.py
