+++ .github/scripts/modules_test.sh
@@ -139,6 +139,9 @@ for f in spyder/*/*/*/*/*.py; do
     if [[ $f == spyder/plugins/editor/widgets/codeeditor/__init__.py ]]; then
         continue
     fi
+    if [[ $f == spyder/plugins/editor/widgets/editorstack/__init__.py ]]; then
+        continue
+    fi
     python "$f"
     if [ $? -ne 0 ]; then
         exit 1
