+++ spyder/plugins/editor/widgets/tests/test_editorsplitter.py
@@ -5,7 +5,7 @@
 #
 
 """
-Tests for EditorSplitter class in editor.py
+Tests for EditorSplitter class in splitter.py
 """
 
 # Standard library imports
@@ -21,7 +21,8 @@ import pytest
 from qtpy.QtCore import Qt
 
 # Local imports
-from spyder.plugins.editor.widgets.editor import EditorStack, EditorSplitter
+from spyder.plugins.editor.widgets.editorstack import EditorStack
+from spyder.plugins.editor.widgets.splitter import EditorSplitter
 
 
 # ---- Qt Test Fixtures
